Decompose monolithic main.rs into 17-file component architecture
Refactor the 1,906-line crates/intrada-web/src/main.rs into a
three-layer module structure: components/, views/, and root modules.
This is a pure code-move refactoring with no logic changes.

New structure:
- types.rs: SharedCore type alias, ViewState enum
- helpers.rs: parse_tags, parse_tempo, parse_tempo_display
- validation.rs: validate_piece_form, validate_exercise_form
- data.rs: create_stub_data, SAMPLE_PIECES
- core_bridge.rs: process_effects
- app.rs: App root component
- components/: FormFieldError, LibraryItemCard
- views/: LibraryListView, DetailView, AddPieceForm,
          AddExerciseForm, EditPieceForm, EditExerciseForm

All 82 workspace tests pass, zero clippy warnings, WASM builds.
Largest file is 281 lines (< 300 limit), 4.3% line overhead.
