+++ CLAUDE.md
@@ -8,6 +8,8 @@ Auto-generated from all feature plans. Last updated: 2026-02-08
 - N/A (stub data in-memory; no browser persistence) (003-leptos-app-mvp)
 - Rust stable (1.75+, 2021 edition) + leptos 0.8.x (CSR), crux_core 0.17.0-rc2, tailwindcss v4 (standalone CLI), trunk 0.21.x, console_error_panic_hook, wasm-bindgen, ulid, chrono (004-library-detail-editing)
 - N/A (in-memory stub data; no persistence) (004-library-detail-editing)
+- Rust stable (1.75+, 2021 edition) + leptos 0.7 (CSR), crux_core 0.17.0-rc2 (workspace), send_wrapper 0.6, wasm-bindgen, console_error_panic_hook (005-component-architecture)
+- N/A (stub data in-memory; no persistence changes) (005-component-architecture)
 
 - Rust stable (1.75+, 2021 edition) + rusqlite (bundled), clap 4.5 (derive), ulid, serde, thiserror, anyhow, chrono (001-music-library)
 
@@ -32,9 +34,9 @@ cargo clippy
 Rust stable (1.75+, 2021 edition): Follow standard conventions
 
 ## Recent Changes
+- 005-component-architecture: Added Rust stable (1.75+, 2021 edition) + leptos 0.7 (CSR), crux_core 0.17.0-rc2 (workspace), send_wrapper 0.6, wasm-bindgen, console_error_panic_hook
 - 004-library-detail-editing: Added Rust stable (1.75+, 2021 edition) + leptos 0.8.x (CSR), crux_core 0.17.0-rc2, tailwindcss v4 (standalone CLI), trunk 0.21.x, console_error_panic_hook, wasm-bindgen, ulid, chrono
 - 003-leptos-app-mvp: Added Rust stable (1.75+, 2021 edition) — same as existing workspace + leptos 0.8.x (csr), crux_core 0.17.0-rc2 (workspace), tailwindcss v4 (standalone CLI), trunk 0.21.x, console_error_panic_hook, wasm-bindgen
-- 002-ci-cd: Added Rust stable (currently 1.89.0, project minimum 1.75+, 2021 edition) + GitHub Actions, `dtolnay/rust-toolchain@stable`, `Swatinem/rust-cache@v2`
 
 
 <!-- MANUAL ADDITIONS START -->
