+++ crates/intrada-web/src/views/mod.rs
@@ -1,0 +1,13 @@
+pub mod add_exercise;
+pub mod add_piece;
+pub mod detail;
+pub mod edit_exercise;
+pub mod edit_piece;
+pub mod library_list;
+
+pub use add_exercise::AddExerciseForm;
+pub use add_piece::AddPieceForm;
+pub use detail::DetailView;
+pub use edit_exercise::EditExerciseForm;
+pub use edit_piece::EditPieceForm;
+pub use library_list::LibraryListView;
